// src/data/node-types/racing-events/types/ApiRacingEventHasImageRelationship.ts
import type {ApiImageNode} from "../../images/types/ApiImageNode"

export type ApiRacingEventHasImageRelationship = {
    data: {
        relationship_id: number
        relationship_name: 'has-image'
        partner_node: {
            node_type: 'image'
            data: ApiImageNode['attributes']
        }
        created_at: string
        updated_at: string
    }
}